chore: refactors staking banners to be component based
