+++ apps/ledger-live-desktop/src/renderer/families/ethereum/StakeBanner.tsx
@@ -1,23 +1,30 @@
-import { AccountBannerState } from "@ledgerhq/live-common/lib/families/ethereum/banner";
 import { Account } from "@ledgerhq/types-live";
+import { useHistory } from "react-router";
+import { useTranslation } from "react-i18next";
+import useFeature from "@ledgerhq/live-common/featureFlags/useFeature";
+
+import { getAccountBannerState as getEthereumBannerState } from "@ledgerhq/live-common/families/ethereum/banner";
+import { AccountBanner } from "~/renderer/screens/account/AccountBanner";
 import { track } from "~/renderer/analytics/segment";
 import { stakeDefaultTrack } from "~/renderer/screens/stake/constants";
-import { Hooks } from "~/renderer/screens/account/useGetBannerProps";
-import { BannerProps } from "~/renderer/screens/account/AccountBanner";
+import React from "react";
+import { StakeAccountBannerParams } from "~/renderer/screens/account/types";
 
 const kilnAppId = "kiln";
 const lidoAppId = "lido";
 
-const getAccountBannerProps = (
-  state: AccountBannerState,
-  account: Account,
-  { t, ethStakingProviders, history, stakeAccountBannerParams }: Hooks,
-): BannerProps | { display: false } => {
+export const EthereumStakeBanner: React.FC<{ account: Account }> = ({ account }) => {
+  const history = useHistory();
+  const { t } = useTranslation();
+  const stakeAccountBanner = useFeature("stakeAccountBanner");
+  const ethStakingProviders = useFeature("ethStakingProviders");
+  const stakeAccountBannerParams: StakeAccountBannerParams | null =
+    stakeAccountBanner?.params ?? null;
+  const state = getEthereumBannerState(account);
   const { stakeProvider } = state;
 
-  if (stakeProvider === "lido" && !stakeAccountBannerParams?.eth?.lido) return { display: false };
-  if (stakeProvider === "kiln" && !stakeAccountBannerParams?.eth?.kiln) return { display: false };
-
+  if (stakeProvider === "lido" && !stakeAccountBannerParams?.eth?.lido) return null;
+  if (stakeProvider === "kiln" && !stakeAccountBannerParams?.eth?.kiln) return null;
   const stakeOnClick = (providerLiveAppId: string) => {
     track("button_clicked", {
       ...stakeDefaultTrack,
@@ -41,7 +48,6 @@ const getAccountBannerProps = (
     stakeProvider === "kiln"
       ? t("account.banner.delegation.ethereum.kiln.description")
       : t("account.banner.delegation.ethereum.lido.description");
-
   const cta =
     stakeProvider === "kiln"
       ? t("account.banner.delegation.ethereum.kiln.cta")
@@ -54,15 +60,17 @@ const getAccountBannerProps = (
     (provider: { liveAppId: string }) => provider.liveAppId === stakeProvider,
   );
 
-  return {
-    title,
-    description,
-    cta,
-    onClick,
-    display,
-    linkText: t("account.banner.delegation.linkText"),
-    linkUrl: "https://www.ledger.com/staking-ethereum",
-  };
-};
+  if (!display) return null;
 
-export { getAccountBannerProps };
+  return (
+    <AccountBanner
+      title={title}
+      description={description}
+      cta={cta}
+      onClick={onClick}
+      display={true}
+      linkText={t("account.banner.delegation.linkText")}
+      linkUrl={"https://www.ledger.com/staking-ethereum"}
+    />
+  );
+};
