+++ apps/ledger-live-desktop/src/renderer/screens/account/index.jsx
@@ -4,24 +4,22 @@ import React, { useCallback } from "react";
 import { compose } from "redux";
 import { connect, useSelector } from "react-redux";
 import { withTranslation } from "react-i18next";
-import type { TFunction } from "react-i18next";
 import { Redirect } from "react-router";
-import type { AccountLike, Account } from "@ledgerhq/types-live";
 import { SyncOneAccountOnMount } from "@ledgerhq/live-common/bridge/react/index";
 import { findCompoundToken } from "@ledgerhq/live-common/currencies/index";
 import { isNFTActive } from "@ledgerhq/live-common/nft/support";
 import { getCurrencyColor } from "~/renderer/getCurrencyColor";
 import { accountSelector } from "~/renderer/reducers/accounts";
 import {
-  isAccountEmpty,
+  findSubAccountById,
   getAccountCurrency,
   getMainAccount,
-  findSubAccountById,
+  isAccountEmpty,
 } from "@ledgerhq/live-common/account/index";
 import { setCountervalueFirst } from "~/renderer/actions/settings";
 import {
-  hiddenNftCollectionsSelector,
   countervalueFirstSelector,
+  hiddenNftCollectionsSelector,
 } from "~/renderer/reducers/settings";
 
 import TrackPage from "~/renderer/analytics/TrackPage";
@@ -39,8 +37,9 @@ import EmptyStateAccount from "./EmptyStateAccount";
 import TokensList from "./TokensList";
 import CompoundBodyHeader from "~/renderer/screens/lend/Account/AccountBodyHeader";
 import useCompoundAccountEnabled from "~/renderer/screens/lend/useCompoundAccountEnabled";
-import { AccountBanner } from "./AccountBanner";
-import { useGetBannerProps } from "./useGetBannerProps";
+import { AccountStakeBanner } from "~/renderer/screens/account/AccountStakeBanner";
+import type { TFunction } from "react-i18next";
+import type { AccountLike, Account } from "@ledgerhq/types-live";
 
 const mapStateToProps = (
   state,
@@ -92,7 +91,6 @@ const AccountPage = ({
     : null;
   const bgColor = useTheme("colors.palette.background.paper");
   const isCompoundEnabled = useCompoundAccountEnabled(account, parentAccount);
-  const banner = useGetBannerProps(account);
 
   const hiddenNftCollections = useSelector(hiddenNftCollectionsSelector);
   const filterOperations = useCallback(
@@ -159,7 +157,7 @@ const AccountPage = ({
               ctoken={ctoken}
             />
           </Box>
-          <AccountBanner {...banner} />
+          <AccountStakeBanner account={account} />
           {AccountBodyHeader ? (
             <AccountBodyHeader account={account} parentAccount={parentAccount} />
           ) : null}
