+++ apps/ledger-live-desktop/src/renderer/screens/account/types.ts
@@ -1,0 +1,18 @@
+export type StakeAccountBannerParams = {
+  solana?: {
+    redelegate: boolean;
+    delegate: boolean;
+  };
+  eth?: {
+    kiln: boolean;
+    lido: boolean;
+  };
+  osmos?: {
+    redelegate: boolean;
+    delegate: boolean;
+  };
+  cosmos?: {
+    redelegate: boolean;
+    delegate: boolean;
+  };
+};
